Added 'moved' files back to source control, committing those now

+++ Duo Play/TournamentSettingsView.swift
@@ -15,7 +15,10 @@ import Crashlytics
 // controlling flow between pool play and bracket or Bracket only.
 class TournamentSettingsView: UIViewController {
     
-    @IBOutlet weak var tournamentNameTextField: UITextField!
+	@IBOutlet weak var challongeLinkLabel: UILabel!
+	@IBOutlet weak var isPublicSwitch: UISwitch!
+	@IBOutlet weak var isOnlineSwitch: UISwitch!
+	@IBOutlet weak var tournamentNameTextField: UITextField!
     @IBOutlet weak var isQuickReportSwitch: UISwitch!
     @IBOutlet weak var isBracketOnlySwitch: UISwitch!
     @IBOutlet weak var playersPerPoolSegementedControl: UISegmentedControl!
@@ -50,7 +53,7 @@ class TournamentSettingsView: UIViewController {
 			playersPerPoolLabel.isHidden = true
 		}
 		
-		if tournament.progress_meter > 0 {
+		if tournament.progress_meter > 0 || tournament.isReadOnly {
 			// tournament has begun, don't let settings be editable
 			bracketOnlyButton.isEnabled = false
 			poolPlayAndBracketButton.isEnabled = false
@@ -59,18 +62,32 @@ class TournamentSettingsView: UIViewController {
 		}
 		
 		tournamentNameTextField.text = tournament.name
+		challongeLinkLabel.text = "Challonge Link : " + tournament.full_challonge_url
+		
+		// tournament is read only, let's hide everything!
+		if tournament.isReadOnly {
+			bracketOnlyButton.isHidden = true
+			poolPlayAndBracketButton.isHidden = true
+			isOnlineSwitch.isHidden = true
+			isPublicSwitch.isHidden = true
+			tournamentNameTextField.isEnabled = false
+			advanceButton.setTitle("Next", for: .normal)
+		}
     }
 	
 	@IBAction func deleteButton(_ sender: UIButton) {
 		// show safety dialog first.
 		let alert = UIAlertController(title: "Delete Tourament",
-									  message: "Everything Will Be Deleted!", preferredStyle: .alert)
+									  message: "Everything will be deleted!", preferredStyle: .alert)
 		
 		let action = UIAlertAction(title: "Delete", style: .destructive) { (alertAction) in
-			self.deleteTournament()
-			
-			Answers.logCustomEvent(withName: "Tournament Deleted",
-								   customAttributes: [:])
+			if self.tournament.isReadOnly {
+				self.deleteTournament(deleteOnline: false)
+			} else if self.tournament.password.count > 0 {
+				self.showPasswordAlert()
+			} else {
+				self.deleteTournament(deleteOnline: true)
+			}
 		}
 		
 		alert.addAction(action)
@@ -83,7 +100,7 @@ class TournamentSettingsView: UIViewController {
 		present(alert, animated: true, completion: nil)
 	}
 	
-	func deleteTournament() {
+	func deleteTournament(deleteOnline: Bool) {
 		var count = 0
 		if realm.isInWriteTransaction {
 			count = realm.objects(Tournament.self).filter("id = \(tournament.id)").count
@@ -94,8 +111,11 @@ class TournamentSettingsView: UIViewController {
 		}
 		
 		if count == 0 { return }
-		let tournamentDAO = TournamentDAO()
-		tournamentDAO.deleteOnlineTournament(tournament: tournament)
+		
+		if deleteOnline {
+			let tournamentDAO = TournamentDAO()
+			tournamentDAO.deleteOnlineTournament(tournament: tournament)
+		}
 		
 		try! realm.write {
 			for team in tournament.teamList {
@@ -119,7 +139,58 @@ class TournamentSettingsView: UIViewController {
 			realm.delete(tournament)
 		}
 		
+		Answers.logCustomEvent(withName: "Tournament Deleted",
+							   customAttributes: [:])
 		navigationController?.popViewController(animated: true)
+	}
+	
+	func showPasswordAlert() {
+		let alert = UIAlertController(title: "Password",
+									  message: "Please enter password to delete this tournament online.", preferredStyle: .alert)
+		
+		let submit = UIAlertAction(title: "Submit", style: .default) { (alertAction) in
+			_ = alert.textFields![0] as UITextField
+			let pw = alert.textFields![0].text!
+			let password = self.tournament.password
+			
+			if pw == password {
+				self.deleteTournament(deleteOnline: true)
+			} else {
+				self.showPasswordIncorrectAlert()
+			}
+		}
+		
+		alert.addAction(submit)
+		
+		alert.addTextField { (textField) in
+			textField.placeholder = "Password"
+			textField.borderStyle = UITextBorderStyle.roundedRect
+		}
+		
+		
+		alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action: UIAlertAction!) in
+			// cancel
+			return
+		}))
+		
+		present(alert, animated: true, completion: nil)
+	}
+	
+	func showPasswordIncorrectAlert() {
+		let alert = UIAlertController(title: "Password Incorrect",
+									  message: "Do you wish to delete this tournament from your device?", preferredStyle: .alert)
+		
+		alert.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { (action: UIAlertAction!) in
+			// delete
+			self.deleteTournament(deleteOnline: false)
+		}))
+		
+		alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: { (action: UIAlertAction!) in
+			// cancel
+			return
+		}))
+		
+		present(alert, animated: true, completion: nil)
 	}
 	
 	@IBAction func poolPlayAndBracketButton(_ sender: UIButton) {
@@ -150,18 +221,28 @@ class TournamentSettingsView: UIViewController {
         TournamentController.IS_QUICK_REPORT = false //isQuickReportSwitch.isOn
         
         try! realm.write {
-            tournament.isQuickReport = false //isQuickReportSwitch.isOn
-            tournament.playersPerPool = playersPerPoolSegementedControl.selectedSegmentIndex + 6
-            tournament.name = (tournamentNameTextField.text?.count.magnitude)! > 0 ?
-                tournamentNameTextField.text! :
-                tournament.name
+			if !tournament.isReadOnly {
+				tournament.isOnline = isOnlineSwitch.isOn
+				tournament.isPrivate = !isPublicSwitch.isOn
+				tournament.isQuickReport = false //isQuickReportSwitch.isOn
+				tournament.playersPerPool = playersPerPoolSegementedControl.selectedSegmentIndex + 6
+				tournament.name = (tournamentNameTextField.text?.count.magnitude)! > 0 ?
+					tournamentNameTextField.text! :
+					tournament.name
+				
+				let tournamentDao = TournamentDAO()
+				tournamentDao.addOnlineTournament(tournament: tournament)
+				
+				//		let challongeAPI = ChallongeAPI()
+				//		challongeAPI.createTournament(tournament: tournament)
+			}
         }
 		
 		Answers.logCustomEvent(withName: "Tournament Settings Saved",
-							   customAttributes: [
-								"isPoolPlay": String(tournament.isPoolPlay)])
-		
-		let tournamentDao = TournamentDAO()
-		tournamentDao.updateOnlineTournament(tournament: tournament)
+		   customAttributes: [
+			"isPoolPlay": String(tournament.isPoolPlay),
+			"isOnline": String(tournament.isOnline),
+			"isPublic": String(!tournament.isPrivate),
+			"isPassword": String(tournament.password.count > 0)])
     }
 }
